+++ src/authorization/PermissionReader.ts
@@ -1,26 +1,22 @@
 import type { CredentialSet } from '../authentication/Credentials';
-import type { ResourceIdentifier } from '../http/representation/ResourceIdentifier';
 import { AsyncHandler } from '../util/handlers/AsyncHandler';
-import type { AccessMode, PermissionSet } from './permissions/Permissions';
+import type { AccessMap, PermissionMap } from './permissions/Permissions';
 
 export interface PermissionReaderInput {
   /**
-   * Credentials of the entity that wants to use the resource.
+   * Credentials of the entity requesting access to resources.
    */
   credentials: CredentialSet;
   /**
-   * Identifier of the resource that will be read/modified.
+   * For each credential, the reader will check which of the given per-resource access modes are available.
+   * However, non-exhaustive information about other access modes and resources can still be returned.
    */
-  identifier: ResourceIdentifier;
-  /**
-   * This is the minimum set of access modes the output needs to contain,
-   * allowing the handler to limit its search space to this set.
-   * However, non-exhaustive information about other access modes can still be returned.
-   */
-  modes: Set<AccessMode>;
+  requestedModes: AccessMap;
 }
 
 /**
  * Discovers the permissions of the given credentials on the given identifier.
+ * In case the reader finds no permission for the requested identifiers and credentials
+ * it can return an empty or incomplete map.
  */
-export abstract class PermissionReader extends AsyncHandler<PermissionReaderInput, PermissionSet> {}
+export abstract class PermissionReader extends AsyncHandler<PermissionReaderInput, PermissionMap> {}