+++ src/storage/DataAccessorBasedStore.ts
@@ -19,6 +19,8 @@ import { NotFoundHttpError } from '../util/errors/NotFoundHttpError';
 import { NotImplementedHttpError } from '../util/errors/NotImplementedHttpError';
 import { PreconditionFailedHttpError } from '../util/errors/PreconditionFailedHttpError';
 import type { IdentifierStrategy } from '../util/identifiers/IdentifierStrategy';
+import { concat } from '../util/IterableUtil';
+import { IdentifierMap } from '../util/map/IdentifierMap';
 import {
   ensureTrailingSlash,
   isContainerIdentifier,
@@ -38,11 +40,13 @@ import {
   SOLID_META,
   PREFERRED_PREFIX_TERM,
   CONTENT_TYPE_TERM,
+  SOLID_AS,
+  AS,
 } from '../util/Vocabularies';
 import type { DataAccessor } from './accessors/DataAccessor';
 import type { Conditions } from './Conditions';
 import type { RepresentationConverter } from './conversion/RepresentationConverter';
-import type { ResourceStore } from './ResourceStore';
+import type { ResourceStore, ChangeMap } from './ResourceStore';
 
 /**
  * ResourceStore which uses a DataAccessor for backend access.
@@ -159,7 +163,7 @@ export class DataAccessorBasedStore implements ResourceStore {
   }
 
   public async addResource(container: ResourceIdentifier, representation: Representation, conditions?: Conditions):
-  Promise<ResourceIdentifier> {
+  Promise<ChangeMap> {
     this.validateIdentifier(container);
 
     const parentMetadata = await this.getSafeNormalizedMetadata(container);
@@ -195,13 +199,11 @@ export class DataAccessorBasedStore implements ResourceStore {
     }
 
     // Write the data. New containers should never be made for a POST request.
-    await this.writeData(newID, representation, isContainer, false, false);
-
-    return newID;
+    return this.writeData(newID, representation, isContainer, false, false);
   }
 
   public async setRepresentation(identifier: ResourceIdentifier, representation: Representation,
-    conditions?: Conditions): Promise<ResourceIdentifier[]> {
+    conditions?: Conditions): Promise<ChangeMap> {
     this.validateIdentifier(identifier);
 
     // Check if the resource already exists
@@ -241,7 +243,7 @@ export class DataAccessorBasedStore implements ResourceStore {
   }
 
   public async modifyResource(identifier: ResourceIdentifier, patch: Patch,
-    conditions?: Conditions): Promise<ResourceIdentifier[]> {
+    conditions?: Conditions): Promise<never> {
     if (conditions) {
       let metadata: RepresentationMetadata | undefined;
       try {
@@ -258,7 +260,7 @@ export class DataAccessorBasedStore implements ResourceStore {
     throw new NotImplementedHttpError('Patches are not supported by the default store.');
   }
 
-  public async deleteResource(identifier: ResourceIdentifier, conditions?: Conditions): Promise<ResourceIdentifier[]> {
+  public async deleteResource(identifier: ResourceIdentifier, conditions?: Conditions): Promise<ChangeMap> {
     this.validateIdentifier(identifier);
 
     // https://github.com/CommunitySolidServer/CommunitySolidServer/issues/1027#issuecomment-988664970
@@ -298,22 +300,25 @@ export class DataAccessorBasedStore implements ResourceStore {
     // Solid, §5.4: "When a contained resource is deleted,
     // the server MUST also delete the associated auxiliary resources"
     // https://solid.github.io/specification/protocol#deleting-resources
-    const deleted = [ identifier ];
+    const changes: ChangeMap = new IdentifierMap();
     if (!this.auxiliaryStrategy.isAuxiliaryIdentifier(identifier)) {
       const auxiliaries = this.auxiliaryStrategy.getAuxiliaryIdentifiers(identifier);
-      deleted.push(...await this.safelyDeleteAuxiliaryResources(auxiliaries));
+      for (const deletedId of await this.safelyDeleteAuxiliaryResources(auxiliaries)) {
+        this.addActivityMetadata(changes, deletedId, AS.terms.Delete);
+      }
     }
 
     if (!this.identifierStrategy.isRootContainer(identifier)) {
       const container = this.identifierStrategy.getParentContainer(identifier);
-      deleted.push(container);
+      this.addActivityMetadata(changes, container, AS.terms.Update);
 
       // Update modified date of parent
       await this.updateContainerModifiedDate(container);
     }
 
     await this.accessor.deleteResource(identifier);
-    return deleted;
+    this.addActivityMetadata(changes, identifier, AS.terms.Delete);
+    return changes;
   }
 
   /**
@@ -387,7 +392,7 @@ export class DataAccessorBasedStore implements ResourceStore {
    * @returns Identifiers of resources that were possibly modified.
    */
   protected async writeMetadata(identifier: ResourceIdentifier, representation: Representation):
-  Promise<ResourceIdentifier[]> {
+  Promise<ChangeMap> {
     const subjectIdentifier = this.metadataStrategy.getSubjectIdentifier(identifier);
 
     // Cannot create metadata without a corresponding resource
@@ -402,6 +407,8 @@ export class DataAccessorBasedStore implements ResourceStore {
         'Not allowed to create metadata resources on a metadata resource.',
       );
     }
+
+    const changes: ChangeMap = new IdentifierMap();
 
     const metadata = new RepresentationMetadata(subjectIdentifier);
     let rdf = representation;
@@ -419,7 +426,8 @@ export class DataAccessorBasedStore implements ResourceStore {
     this.removeResponseMetadata(metadata);
     metadata.removeAll(CONTENT_TYPE_TERM);
     await this.accessor.writeMetadata(subjectIdentifier, metadata);
-    return [ identifier ];
+    this.addActivityMetadata(changes, identifier, AS.terms.Create);
+    return changes;
   }
 
   /**
@@ -434,7 +442,7 @@ export class DataAccessorBasedStore implements ResourceStore {
    * @returns Identifiers of resources that were possibly modified.
    */
   protected async writeData(identifier: ResourceIdentifier, representation: Representation, isContainer: boolean,
-    createContainers: boolean, exists: boolean): Promise<ResourceIdentifier[]> {
+    createContainers: boolean, exists: boolean): Promise<ChangeMap> {
     // Make sure the metadata has the correct identifier and correct type quads
     // Need to do this before handling container data to have the correct identifier
     representation.metadata.identifier = DataFactory.namedNode(identifier.path);
@@ -457,18 +465,21 @@ export class DataAccessorBasedStore implements ResourceStore {
     // Solid, §5.3: "Servers MUST create intermediate containers and include corresponding containment triples
     // in container representations derived from the URI path component of PUT and PATCH requests."
     // https://solid.github.io/specification/protocol#writing-resources
-    const modified = [];
+    let changes: ChangeMap = new IdentifierMap();
     if (!this.identifierStrategy.isRootContainer(identifier) && !exists) {
-      const container = this.identifierStrategy.getParentContainer(identifier);
-      if (!createContainers) {
-        modified.push(container);
-      } else {
-        const created = await this.createRecursiveContainers(container);
-        modified.push(...created.length === 0 ? [ container ] : created);
+      const parent = this.identifierStrategy.getParentContainer(identifier);
+
+      if (createContainers) {
+        changes = await this.createRecursiveContainers(parent);
+      }
+
+      // No changes means the parent container exists and will be updated
+      if (changes.size === 0) {
+        this.addActivityMetadata(changes, parent, AS.terms.Update);
       }
 
       // Parent container is also modified
-      await this.updateContainerModifiedDate(container);
+      await this.updateContainerModifiedDate(parent);
     }
 
     // Remove all generated metadata to prevent it from being stored permanently
@@ -478,7 +489,8 @@ export class DataAccessorBasedStore implements ResourceStore {
       this.accessor.writeContainer(identifier, representation.metadata) :
       this.accessor.writeDocument(identifier, representation.data, representation.metadata));
 
-    return [ ...modified, identifier ];
+    this.addActivityMetadata(changes, identifier, exists ? AS.terms.Update : AS.terms.Create);
+    return changes;
   }
 
   /**
@@ -654,11 +666,11 @@ export class DataAccessorBasedStore implements ResourceStore {
    * Will throw errors if the identifier of the last existing "container" corresponds to an existing document.
    * @param container - Identifier of the container which will need to exist.
    */
-  protected async createRecursiveContainers(container: ResourceIdentifier): Promise<ResourceIdentifier[]> {
+  protected async createRecursiveContainers(container: ResourceIdentifier): Promise<ChangeMap> {
     // Verify whether the container already exists
     try {
       const metadata = await this.getNormalizedMetadata(container);
-      // See #480
+      // See https://github.com/CommunitySolidServer/CommunitySolidServer/issues/480
       // Solid, §3.1: "If two URIs differ only in the trailing slash, and the server has associated a resource with
       // one of them, then the other URI MUST NOT correspond to another resource. Instead, the server MAY respond to
       // requests for the latter URI with a 301 redirect to the former."
@@ -666,7 +678,7 @@ export class DataAccessorBasedStore implements ResourceStore {
       if (!isContainerPath(metadata.identifier.value)) {
         throw new ForbiddenHttpError(`Creating container ${container.path} conflicts with an existing resource.`);
       }
-      return [];
+      return new IdentifierMap();
     } catch (error: unknown) {
       if (!NotFoundHttpError.isInstance(error)) {
         throw error;
@@ -674,10 +686,21 @@ export class DataAccessorBasedStore implements ResourceStore {
     }
 
     // Create the container, starting with its parent
-    const ancestors = this.identifierStrategy.isRootContainer(container) ?
-      [] :
+    const ancestors: ChangeMap = this.identifierStrategy.isRootContainer(container) ?
+      new IdentifierMap() :
       await this.createRecursiveContainers(this.identifierStrategy.getParentContainer(container));
-    await this.writeData(container, new BasicRepresentation([], container), true, false, false);
-    return [ ...ancestors, container ];
+    const changes = await this.writeData(container, new BasicRepresentation([], container), true, false, false);
+
+    return new IdentifierMap(concat([ changes, ancestors ]));
+  }
+
+  /**
+   * Generates activity metadata for a resource and adds it to the {@link ChangeMap}
+   * @param map - ChangeMap to update.
+   * @param id - Identifier of the resource being changed.
+   * @param activity - Which activity is taking place.
+   */
+  private addActivityMetadata(map: ChangeMap, id: ResourceIdentifier, activity: NamedNode): void {
+    map.set(id, new RepresentationMetadata(id, { [SOLID_AS.Activity]: activity }));
   }
 }