+++ RELEASE_NOTES.md
@@ -40,7 +40,22 @@ The following changes are relevant for v4 custom configs that replaced certain f
 - The `IdentityProviderFactory` inputs have been extended.
   - `/identity/handler/provider-factory/identity.json`
 - LDP components have slightly changed so the preference parser is in a separate config file.
-  - `/config/ldp/handler/*`
+  - `/ldp/handler/*`
+- Restructured the init configs.
+  - `/app/init/base/init.json`
+  - `/app/main/default.json`
+- Added lock cleanup on server start (and updated existing finalization).
+  - `/util/resource-locker/file.json`
+  - `/util/resource-locker/redis.json`
+- Updated finalizers.
+  - `/app/identity/handler/account-store/default.json` 
+  - `/identity/ownership/token.json`
+  - `/ldp/authorization/readers/access-checkers/agent-group.json`
+  - `/ldp/handler/*`
+- `IntermediateModesExtractor` has been added to the `ModesExtractors`
+  - `/ldp/modes/default.json`
+- The `PermissionReader` structure has changed to be more consistent.
+  - `/ldp/authorization/*`
 - The `PutOperationHandler` constructor now has an extra argument: `metadataStrategy`, resulting in a change in `config/ldp/handler/components/operation-handler.json`
 - A new key-value pair was added to `config/ldp/metadata-writer/writers/link-rel.json`, which allows generating the link to the description resource
 - The `DataAccessorBasedStore` constructor now has two new arguments: `metadataStrategy` and `converter`. As a result following configuration files are changed:
@@ -58,11 +73,20 @@ These changes are relevant if you wrote custom modules for the server that depen
 - Both `TemplateEngine` implementations now take a `baseUrl` parameter as input.
 - The `IdentityProviderFactory` and `ConvertingErrorHandler` now additionally take a `PreferenceParser` as input.
 - Error handlers now take the incoming HttpRequest as input instead of just the preferences.
+- Extended the initialization/finalization system:
+  * Introduced `Initializable` interface and `InitializableHandler` wrapper class.
+  * Introduced `Finalizer` abstract class and `FinalizableHandler` wrapper class.
+  * Changed type for `finalizer` attribute in `App` from `Finalizable` to `Finalizer` and updated the calling code in `App.stop()`.
+  * Removed the now obsolete `ParallelFinalizer` util class.
+- Added a lock cleanup on initialize for lock implementations `RedisLocker` and `FileSystemResourceLocker`.
+- `ResourceStore` functions that change a resource now return metadata for every changed resource.
+- All permission related interfaces have changed to support permissions over multiple identifiers.
+- `IdentifierStrategy` has a new `contains` method.
 - `DataAccessor` interface is changed. There is now a new method called `writeMetadata`.
 
 A new interface `SingleThreaded` has been added. This empty interface can be implemented to mark a component as not-threadsafe. When the CSS starts in multithreaded mode, it will error and halt if any SingleThreaded components are instantiated.
 
-## V4.0.1
+## v4.0.1
 Freezes the `oidc-provider` dependency to prevent a potential issue with the solid authn client
 as described in https://github.com/inrupt/solid-client-authn-js/issues/2103.